feat(api)!: `this.$end()` and `this.$post()`
Since the commit `744abbc` allowed `this` in `Action` to access other
`Actions`, the pre-existing methods of `this` in Action which named `post` and
`end` need to have their names changed to prevent potential conflicts with the
name of `Action`.

BREAKING CHANGE: use `this.$end()` and `this.$post()` instead of
`this.end()` and `this.post()` in the `Action`.