fix: wrong receiving port
Due to commit `744abbc`, which extends `this` in `Action`, and commit `f7ae71c`, which binds the same `this` to all methods in `Actions`, the unique identifier that can be obtained in `this.$post()` or `this.$end()`, when called within each `Action`, is the same. This causes them to always post the message to the same receiving port.

To ensure the unique identifier obtained in `this.$post()` or `this.$end()` is correct, the methods in `Actions` should not be replaced. They should be called by `apply()` to dynamically assign `this`.