feat!: automatically identify transferable objects
In addition to using the `transfer` array to transfer transferable
objects, they can also be automatically identified.

BREAKING CHANGE: if the data being posted contains transferable objects
and the `transfer` array is not provided, compared to previous versions,
the default behavior is now to transfer all the `transferable objects`
contained in the data.