fix!: inappropriate identification of `transferable objects`
Recursive identification of `transferable objects` from complicated
objects may cause maximum call stack exceeded, hence the automatic
identification of `transferable objects` is no longer the default.

BREAKING CHANGE: when passing data, if the `transfer` option is not
specified, or when using the return value of an `Action` to respond
with data,the default behavior is to not transfer anything rather than
automatically identifying and transferring the `transferable objects`.