fix: some errors with the revocation of `Worker Proxy`
When a `Worker Proxy` is revoked, the target data referenced by its
children or adoptedChildren will also be cleared. However, the children
and adoptedChildren themselves are not revoked. If they are handled, an
error will be thrown in `Worker` indicating that the referenced data has
been cleared, but it will not be able to locate the specific line of
code where they were handled.

In this commit, when a `Worker Proxy` is revoked, its children and
adoptedChildren will also be revoked.